==> scripts_sequencing/9_prey_diversity.Rmd <==
---
title: "Results: Prey Diversity"
author: "M Fisher"
date: "2023-03-26"
output: 
  html_document:
    toc: yes
    toc_float: yes
---

# Description

Summarize the following across sequencing runs:

1. Sample sizes
  a. how many crab were sequenced?
  b. how many crab had prey information? **Table 1**
  c. how large were the crab that were sequenced v trapped? **Table 1**

2. Specificity: what proportion of reads were identified down to species? Genus? etc.

3. Alpha diversity
  a. What was the unique number of taxa identified overall?
  b. What was the unique number of taxa identified per site or month?  **Table 1**
  
3. Beta diversity **Table 2**

4. Frequency of occurrence **Figure 2**



```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message=FALSE)
library(tidyverse)
library(here)
library(ggplot2)
library(ggrepel)
library(cowplot)
library(janitor)
library(magrittr)
library(vegan)
library(stargazer)

# source(here('R','gg_ordiplot_custom.R'))
source(here('R','eDNA_index.R'))
source(here('R','get_unique_taxa.R'))

##### User inputs #####
run.nums <- c(2,3,4,5)
marker  <- 'BF3'
meta_dir <- 'data/metadata'
results_dir <- 'data/results'
writing <- FALSE
```
```{r data}
# metadata from trapping
meta.trap <- read_csv(here(meta_dir,'Willapa Bay EGC Samples - Sample Data.csv'))

# metadata for run
meta.run <- read_csv(here(meta_dir,'runs2-5_BF3_run_metadata.csv'))

# sequencing data file - final manual filtered
dat  <- read_csv(here(results_dir, 'allRuns_BF3_filtered_FINAL_unique_taxa.csv'))

# count of crab dissected by site
site_coords <- read_csv(here(meta_dir,'site_coords.csv'))

set.seed(555)
```

# Sample Size

How many crab were sequenced, per site and month?

```{r}
meta.run %>%
  filter(!is.na(site_month)) %>%
  group_by(site_month) %>%
  summarise(`sequenced`=length(unique(sample)))
```


## n crab per step

```{r fig.height=3, fig.width=5}
site_coords %>% summarise(n_dissected=sum(n_dissected))

meta.run %>%
  filter(!is.na(site_month)) %>%
  summarise(`sequenced`=length(unique(sample)))

dat %>%
  filter(!is.na(site_month)) %>%
  summarise(`with prey`=length(unique(sample))) %>%
  mutate(`with prey`=ifelse(is.na(`with prey`),0,`with prey`))

```

```{r fig.height=3, fig.width=5}
dissected <- site_coords %>% group_by(site_name) %>% summarise(n_dissected=sum(n_dissected)) %>%
  rename(Site=site_name)

sequenced <- meta.run %>%
  filter(!is.na(site_month)) %>%
  separate(site_month, into=c("Site","Month"), sep="-") %>%
  group_by(Site) %>%
  summarise(`sequenced`=length(unique(sample)))

success <- dat %>%
  filter(!is.na(site_month)) %>%
  separate(site_month, into=c("Site","Month"), sep="-") %>%
  group_by(Site) %>%
  summarise(`with prey`=length(unique(sample))) %>%
  mutate(`with prey`=ifelse(is.na(`with prey`),0,`with prey`))

ncrab_bySite <- full_join(sequenced,success,by="Site") %>%
  mutate(`seq with prey` = round(`with prey`/`sequenced`,2)) 
ncrab_bySite
```


```{r}
dissectedM <- site_coords %>% group_by(sampling_month) %>% summarise(n_dissected=sum(n_dissected)) %>%
  rename(Month=sampling_month) %>%
  mutate(Month=ifelse(Month=="July","Jul",
                      ifelse(Month=="September","Sep",Month)))
sequencedM <- meta.run %>%
  filter(!is.na(site_month)) %>%
  separate(site_month, into=c("Site","Month"), sep="-") %>%
  group_by(Month) %>%
  summarise(`sequenced`=length(unique(sample)))

successM <- dat %>%
  filter(!is.na(site_month)) %>%
  separate(site_month, into=c("Site","Month"), sep="-") %>%
  group_by(Month) %>%
  summarise(`with prey`=length(unique(sample))) %>%
  mutate(`with prey`=ifelse(is.na(`with prey`),0,`with prey`))

ncrab_byMonth <- full_join(sequencedM,successM,by="Month") %>%
  mutate(`seq with prey` = round(`with prey`/`sequenced`,2)) 
ncrab_byMonth
```
<br>

## carapace widths

Mean carapace widths (mm): 
```{r}
cw <- meta.trap %>%
  mutate(sequenced=ifelse(Sample_label %in% meta.run$sample, "Sequenced", "Trapped"))

cw %<>% mutate(site=ifelse(grepl("Stackpole",Site_Name), "Stackpole",
                                 ifelse(grepl("Oysterville",Site_Name), "Oysterville",
                                        ifelse(grepl("Nahcotta",Site_Name),"Nahcotta",
                                               ifelse(grepl("Long Beach",Site_Name),"Long Beach", NA)))))

cw %>% group_by(site,sequenced) %>% summarise(mean_cw=mean(CW_mm)) %>% pivot_wider(names_from=sequenced,values_from=mean_cw)
```

Range for **collected** crab?
```{r}
cw %>% group_by(Sex) %>% summarise(n.crab=length(unique(Sample_label)),
                                                                      min.cw=min(CW_mm, na.rm = TRUE),
                                                                      max.cw=max(CW_mm, na.rm = TRUE))
```

Range for **sequenced** crab?
```{r}
cw %>% filter(sequenced=="Sequenced") %>% group_by(Sex) %>% summarise(n.crab=length(unique(Sample_label)),
                                                                      min.cw=min(CW_mm, na.rm = TRUE),
                                                                      max.cw=max(CW_mm, na.rm = TRUE))
```

Range for sequenced crab **with prey**?
```{r}
meta.trap %>%
  mutate(with_prey=ifelse(Sample_label %in% dat$sample, "with Prey", "Trapped")) %>%
  mutate(site=ifelse(grepl("Stackpole",Site_Name), "Stackpole",
                                 ifelse(grepl("Oysterville",Site_Name), "Oysterville",
                                        ifelse(grepl("Nahcotta",Site_Name),"Nahcotta",
                                               ifelse(grepl("Long Beach",Site_Name),"Long Beach", NA))))) %>% 
  filter(with_prey=="with Prey") %>% group_by(Sex) %>% summarise(n.crab=length(unique(Sample_label)),
                                                                      min.cw=min(CW_mm, na.rm = TRUE),
                                                                      max.cw=max(CW_mm, na.rm = TRUE))
```


# Specificity

what **percent** of reads in the final data set were identified down to genus or species for prey species? 

Across all sites
```{r}
total.reads <- sum(dat$nReads)
all.tax.depth <- dat %>%
  group_by(rank) %>%
  summarise(level_reads=sum(nReads),
            prop_level_reads=level_reads/total.reads,
            per_level_reads=prop_level_reads*100)
printvec <- round(pull(all.tax.depth,per_level_reads),3)
names(printvec)<- pull(all.tax.depth,rank)

printvec[printvec %>% order]
```
<br> 

At each site
```{r}
site.tax.depth <- dat %>%
  separate(site_month,into=c("site","month"),sep="-",remove=FALSE) %>%
  group_by(site) %>% mutate(site_reads=sum(nReads)) %>%
  ungroup() %>%
  group_by(site,rank) %>%
  summarise(level_reads=sum(nReads),
            prop_level_reads=level_reads/site_reads,
            per_level_reads=round(prop_level_reads*100,3)) %>%
  ungroup() %>% distinct()

site.tax.depth %>%
  dplyr::select(site,rank,per_level_reads) %>%
  pivot_wider(names_from=rank,values_from=per_level_reads)
```



# Alpha Diversity 

What was the unique number of taxa identified?

## Across sites

All sites
```{r}
length(unique(dat$taxon))
```

*Corrected for potential duplicates at higher taxonomic levels*
```{r}
length(unique(get_unique_taxa(taxa.df=dat, level="all", return.removed=FALSE) %>%
  pull(taxon)))
```

By site type
```{r}
dat %>% separate(site_month, into=c("site","month"), sep="-") %>%
  group_by(site) %>%
  summarise(n=length(unique(taxon)))
```

*Corrected for potential duplicates at higher taxonomic levels*
```{r}
get_unique_taxa(taxa.df = dat %>%  separate(site_month, into=c("site","month"), sep="-"),
                level="site",return.removed=FALSE) %>%
  group_by(site) %>% summarise(n=length(unique(taxon)))
```

Take a quick look at what was removed
```{r eval=FALSE}
View((get_unique_taxa(taxa.df=dat %>%  separate(site_month, into=c("site","month"), sep="-"), level="site", return.removed=TRUE))[[2]])
```


## Per crab stomach

**Figure S2**
```{r fig.height=3, fig.width=4}
crab.tax.breadth <-dat %>% 
  filter(!is.na(site_month)) %>%
  group_by(site_month, sample) %>% 
  summarise(n_taxa=length(unique(species)))

ggplot(crab.tax.breadth, aes(x=n_taxa)) +
  geom_histogram() + 
  labs(x="Number of Prey Taxa in Stomach", y="Number of Crabs") +
  scale_x_continuous(breaks=seq(1,15,by=2)) +
  theme_bw()
```
```{r eval=writing}
png(here('figs','FigS2_alpha-diversity_per-crab.png'),res=300,height=800,width=800)
ggplot(crab.tax.breadth, aes(x=n_taxa)) +
  geom_histogram() + 
  labs(x="Number of Prey Taxa in Stomach", y="Number of Crabs") +
  scale_x_continuous(breaks=seq(1,13,by=2)) +
  theme_bw()
dev.off()
```
<br>

What is in the crab with > 10 species??
```{r}
crab <- (filter(crab.tax.breadth, n_taxa > 10))$sample
crab
dat %>% filter(sample==crab) %>%
  group_by(class,taxon,MiSeqRun) %>% summarise(totalReads=sum(nReads), n_tech=length(unique(tech)))
```
<br>


**Figure S3**

```{r fig.height=3, fig.width=4}
site.tax.breadth <- dat %>% group_by(site_month) %>% summarise(n_taxa=length(unique(taxon)), n_crabs=length(unique(sample))) %>%
  separate(site_month,into=c("site","month"),sep="-",remove=FALSE) %>%
  mutate(site_type=ifelse(site %in% c("Long Beach","Oysterville"),"Clam Bed","Slough"))
site.tax.breadth$site <- factor(site.tax.breadth$site,levels=c("Long Beach","Oysterville","Nahcotta","Stackpole"))

col_palette <- c("#01665e","#a6611a")
ggplot(site.tax.breadth, aes(x=n_crabs,y=n_taxa, col=site_type)) +
  geom_point(size=3) +
  geom_text_repel(aes(label=site_month), size=3, col="black",min.segment.length = 0.01) + 
  labs(x="Number of crabs",y="Number of Taxa") + scale_color_manual(values=col_palette, name="Site Type") +
  xlim(c(0,12)) +
  theme_bw()
```
```{r}
png(here('figs','FigureS3_alphaVsample.png'),res=300,height=1000,width=1400)
ggplot(site.tax.breadth, aes(x=n_crabs,y=n_taxa, col=site_type)) +
  geom_point(size=3) +
  geom_text_repel(aes(label=site_month), size=3, col="black",min.segment.length = 0.01) + 
  labs(x="Number of crabs",y="Number of taxa") + scale_color_manual(values=col_palette, name="Site Type") +
  xlim(c(0,12)) +
  theme_bw()
dev.off()
```


# Beta diversity

What is the $\beta$ diversity for prey taxon? (the ratio between regional and local prey diversity) The greater the similarity in community composition between multiple communities, the lower the value of β-diversity for that set of communities.
```{r}
prey.site.mat <- dat %>%
  separate(site_month,into=c("site","month"),sep="-",remove=FALSE) %>%
  dplyr::select(site, taxon) %>% distinct() %>%
  mutate(presence=1) %>%
  pivot_wider(id_cols="site", names_from="taxon",values_from="presence", values_fill=0)

bdiv <- as.matrix(betadiver(prey.site.mat, method="w")) 
colnames(bdiv) <-c(prey.site.mat$site); rownames(bdiv) <- c(prey.site.mat$site)
bdiv
```


How much changes if we just look at species-level identifications?
```{r}
preySP.site.mat <- dat %>%
  filter(rank=="species") %>%
  separate(site_month,into=c("site","month"),sep="-",remove=FALSE) %>%
  dplyr::select(site, taxon) %>% distinct() %>%
  mutate(presence=1) %>%
  pivot_wider(id_cols="site", names_from="taxon",values_from="presence", values_fill=0)

bdivSP <- as.matrix(betadiver(preySP.site.mat, method="w")) 
colnames(bdivSP) <-c(preySP.site.mat$site); rownames(bdivSP) <- c(preySP.site.mat$site)
bdivSP
```


Oysterville v Long Beach beta diversity is reduced (by 0.04); as is Oysterville and Long Beach v. Nahcotta (by 0.02, 0.06 respectively).





# Frequency of occurrence


## Figure 2

Directly comparable to Fig 1 Cordone et al. by site 

```{r}
total.crab <- length(unique(dat$sample))
plotdat.fo <- dat %>% mutate(plot.phylum=ifelse(is.na(phylum), "Phaeophyceae", phylum)) %>%
  separate(site_month,sep="-",into=c("site","month"), remove=FALSE) %>%
  group_by(taxon,phylum,plot.phylum) %>%
  summarise(ncrab=length(unique(sample))) %>%
  mutate(pcrab=(ncrab/total.crab)*100) %>%
  arrange(plot.phylum)

plotdat.fo %<>% left_join(data.frame(plot.phylum=unique(plotdat.fo$plot.phylum),
                       nphylum=order(unique(plotdat.fo$plot.phylum))),by="plot.phylum") %>%
  unite(plot.phylum, nphylum, col="plot.phylum.num",sep="-", remove=FALSE)

plot2a <- ggplot(plotdat.fo, aes(x=plot.phylum.num,y=pcrab)) +
  geom_boxplot() + geom_hline(aes(yintercept=0)) +
  ylim(c(0,25)) +
  theme_bw() + labs(x="Prey Group", y="Frequency of occurrence (FO%)") + theme(axis.text.x=element_text(angle=60, hjust=1,size=12),
                     legend.text=element_text(size=12),
                     axis.title=element_text(size=13))
plot2a
```


Prey taxa consumed by 4+ crab.


```{r}
plot.taxa <- dat %>% group_by(taxon) %>%
  summarise(ncrab=length(unique(sample))) %>% filter(ncrab>3)

total.crab.site <- dat %>% 
  separate(site_month,sep="-",into=c("site","month"), remove=FALSE) %>%
  group_by(site) %>%
  summarise(site.crab=length(unique(sample)))

plotdat.fo2 <- dat %>% 
  filter(taxon %in% plot.taxa$taxon) %>%
  mutate(plot.phylum=ifelse(is.na(phylum), "Phaeophyceae", phylum)) %>%
  separate(site_month,sep="-",into=c("site","month"), remove=FALSE) %>%
  group_by(site,taxon,plot.phylum) %>%
  summarise(ncrab=length(unique(sample))) %>%
  mutate(pcrab=(ncrab/total.crab*100)) %>%
  left_join(plotdat.fo %>% dplyr::select(plot.phylum,nphylum)) %>%
  arrange(plot.phylum,desc(pcrab))

plotdat.fo2$taxon <- factor(plotdat.fo2$taxon, levels=unique(plotdat.fo2$taxon))


add_groups_df <- data.frame(x=c(1,3.5,6,7,8.5,12,15,16),
                         y=rep(21,8),
                         label=unique(plotdat.fo2$nphylum))

plot2b <- ggplot(plotdat.fo2) +
  geom_col(aes(x=taxon,y=pcrab, fill=site)) + geom_hline(aes(yintercept=0)) +
  geom_vline(aes(xintercept=1.5)) + geom_vline(aes(xintercept=5.5)) + geom_vline(aes(xintercept=6.5)) + 
  geom_vline(aes(xintercept=7.5)) + geom_vline(aes(xintercept=9.5)) + geom_vline(aes(xintercept=14.5)) + 
  geom_vline(aes(xintercept=15.5)) + 
  geom_text(data=add_groups_df,aes(x=x,y=y,label=label),size=5) +
  labs(x="Prey Taxon", y="Frequency of occurrence (FO%)") +
  scale_fill_manual(values=c("#018571","#f8ab5b","#4fc1af","#a6611a"), name="Site") +
  theme_bw() + theme(axis.text.x=element_text(angle=60, hjust=1,size=12, face="italic"),
                     legend.text=element_text(size=12),legend.title=element_text(size=12),
                     axis.title=element_text(size=13))
plot2b
```


```{r eval=writing}
top_row <- plot_grid(plot2a,NULL,ncol=2,rel_widths=c(1,0.22))

# png(here('figs','Fig2_FO.png'),res=300,width=2500,height=3200)
# plot_grid(top_row,plot2b,ncol=1,nrow=2,rel_heights=c(0.8,1), labels=c("a","b"))
# dev.off()

tiff(here('figs','Fig2.tif'),res=300,width=2500,height=3200)
plot_grid(top_row,plot2b,ncol=1,nrow=2,rel_heights=c(0.8,1), labels=c("a","b"))
dev.off()
```


do above, but for all species (not just those that occur in more than 3 crabs)
```{r}
plotdat.fo3 <- dat %>% 
  mutate(plot.phylum=ifelse(is.na(phylum), "Phaeophyceae", phylum)) %>%
  separate(site_month,sep="-",into=c("site","month"), remove=FALSE) %>%
  group_by(site,taxon,plot.phylum) %>%
  summarise(ncrab=length(unique(sample))) %>%
  mutate(pcrab=(ncrab/total.crab*100)) %>%
  left_join(plotdat.fo %>% dplyr::select(plot.phylum,nphylum)) %>%
  arrange(plot.phylum,desc(pcrab))
```


create a fourth data frame that does counts by site type
```{r}
total.crab.site.type <- dat %>% 
  separate(site_month,sep="-",into=c("site","month"), remove=FALSE) %>%
  mutate(site_type=ifelse(site %in% c("Oysterville","Long Beach"), "clam bed","slough")) %>%
  group_by(site_type) %>%
  summarise(site.crab=length(unique(sample)))


plotdat.fo4 <- plotdat.fo3 %>%
  mutate(site_type=ifelse(site %in% c("Oysterville","Long Beach"), "clam bed","slough")) %>%
  group_by(taxon, phylum, plot.phylum, nphylum, site_type) %>%
  summarise(ncrab=sum(ncrab)) %>%
  left_join(total.crab.site.type, by="site_type") %>%
  mutate(pcrab=ncrab/site.crab) %>%
  rename(site=site_type) %>% dplyr::select(-site.crab)
```


Save the FO%
```{r eval=writing}
plotdat.fo %>%
  mutate(site="overall") %>%
  bind_rows(plotdat.fo3) %>%
  bind_rows(plotdat.fo4) %>%
  write_csv(here('data','results','Fig2table_FO.csv'))
```


## Table S4. 

FO at higher taxonomic levels

Counts by genus / family / order / class / phylum
```{r}
tableS6 <- dat %>% group_by(phylum, class, order, family, genus) %>% summarise(n.crab=length(unique(sample))) %>%
  mutate(level="genus") %>%
  bind_rows(
    dat %>% group_by(phylum, class, order, family) %>% summarise(n.crab=length(unique(sample))) %>% mutate(level="family")
  ) %>%
  bind_rows(
    dat %>% group_by(phylum, class, order) %>% summarise(n.crab=length(unique(sample))) %>% mutate(level="order")
  ) %>%
  bind_rows(
    dat %>% group_by(phylum, class) %>% summarise(n.crab=length(unique(sample))) %>% mutate(level="class")
  ) %>%
  bind_rows(
    dat %>% group_by(phylum) %>% summarise(n.crab=length(unique(sample))) %>% mutate(level="phylum")
  )
```


```{r eval=writing}
write_csv(tableS4, here('data','results','S4_Table_higher_level_counts.csv'))
```
